Basic auto advancing on dialogue
There are some limitations here. For example if dialogue spans over
multiple boxes this simply will not work because we can't detect that.
You also need to guess how much time is necessary for text instead of
the text triggering itself when it's over. But this is 'good enough' for
a demo state